[post]slugified fixed
