+++ services/post/src/modules/post/post.service.ts
@@ -1,12 +1,29 @@
 import { Injectable } from '@nestjs/common';
 import { Post, Prisma } from '@prisma/client';
 import { PrismaService } from 'src/prisma/prisma.service';
+import { CreatePostDto } from './post.dto';
+import { slugify } from 'src/common/utils/utils';
 
 @Injectable()
 export class PostService {
   constructor(private readonly prisma: PrismaService) {}
 
-  wirtePost(data: Prisma.PostCreateInput): Promise<Post> {
+  wirtePost(post: CreatePostDto): Promise<Post> {
+    const end: number = post.content.indexOf('\n');
+    let slug: string;
+    if (end == -1 || end > 30) {
+      slug = post.content.slice(0, 30);
+    } else {
+      slug = post.content.slice(0, end);
+    }
+    slug = slugify(slug);
+
+    let data: Prisma.PostCreateInput = {
+      author: post.author,
+      slug,
+      content: post.content,
+    };
+
     return this.prisma.post.create({
       data,
     });
